Refactored and made everything work

+++ src/Sharp48.UserInterfaces/GoogleChromeUI.cs
@@ -52,7 +52,7 @@ namespace Sharp48.UserInterfaces
                         (seed, row) => row.Aggregate(seed, (current, cell) => current + cell?.Value + ","));
                 var grid = Grid.ParseGrid(gridString.Remove(gridString.Length - 1));
                 var gameOver = _driver.ExecuteJavaScript<bool>(@"return GameManager._instance.over");
-                var score = _driver.ExecuteJavaScript<double>(@"return GameManager._instance.score");
+                var score = _driver.ExecuteJavaScript<long>(@"return GameManager._instance.score");
                 return new Game(grid, gameOver, score);
             }
         }
